+++ src/essential-select/essential-select-truncate.pipe.ts
@@ -1,18 +1,18 @@
 import {Pipe, PipeTransform} from '@angular/core';
-import {EssentialSelectComponent} from "./essential-select.component";
+import {EssentialSelectComponent} from './essential-select.component';
 
 @Pipe({
   name: 'essentialSelectTruncate'
 })
-export class EssentialSelectTruncatePipe implements PipeTransform{
+export class EssentialSelectTruncatePipe implements PipeTransform {
   transform(value: string, es: EssentialSelectComponent): string | undefined {
     if (!value) {
       return undefined;
     }
     let answer: string = value;
-    es.findPlaceholderLength(answer);
+    es._findPlaceholderLength(answer);
 
-    if (es.isOpenEditable() && es.hasSearchInput) {
+    if (es._isOpenEditable() && es.hasSearchInput) {
       return value;
     }